% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_functions.R
\name{add_DP_and_NV}
\alias{add_DP_and_NV}
\title{add_DP_and_NV Function}
\usage{
add_DP_and_NV(karyotype, simulated_mutations, coverage, purity)
}
\arguments{
\item{karyotype}{karyotype}

\item{simulated_mutations}{mutations}

\item{coverage}{coverage}

\item{purity}{purity}
}
\value{
get mutation subfunction
}
\description{
This function allows you to simulate the NV and DP for the CN events.
}
