% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnaqc.R
\name{add_drivers_to_segment_plot}
\alias{add_drivers_to_segment_plot}
\title{Add driver mutation annotations to a segment plot}
\usage{
add_drivers_to_segment_plot(x, drivers_list, base_plot)
}
\arguments{
\item{x}{A CNAqc object containing mutation data.}

\item{drivers_list}{A data frame of driver mutations.}

\item{base_plot}{A ggplot object representing the base plot.}
}
\value{
A modified ggplot object with driver mutation annotations.
}
\description{
Add driver mutation annotations to a segment plot
}
