% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr_coordinates_GRCh38}
\alias{chr_coordinates_GRCh38}
\title{GRCh38 coordinates}
\format{
\subsection{\code{chr_coordinates_GRCh38}}{

A data frame:
\describe{
\item{chr}{chromosome identifier}
\item{length}{length of chromosome}
\item{from}{Absolute coordinate of chromosome start}
\item{to}{Absolute coordinate of chromosome end}
\item{centromerStart}{Absolute coordinate of chromosome centromere start}
\item{centromerEnd}{Absolute coordinate of chromosome centromere end}
}
}
}
\source{
.
}
\usage{
chr_coordinates_GRCh38
}
\description{
Reference coordinates (GRCh38)
}
\keyword{datasets}
