% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnaqc.R
\name{convert_name}
\alias{convert_name}
\title{Convert model names to standardized names}
\usage{
convert_name(n)
}
\arguments{
\item{n}{A character string representing the model name.}
}
\value{
A standardized model name as a character string.
}
\description{
Convert model names to standardized names
}
