% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_functions.R
\name{convert_to_granges}
\alias{convert_to_granges}
\title{convert_to_granges Function}
\usage{
convert_to_granges(cna, mutations, metadata)
}
\arguments{
\item{cna}{cna}

\item{mutations}{mutations}

\item{metadata}{metadata}
}
\value{
granges object
}
\description{
convert_to_granges Function
}
