% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_functions.R
\name{convert_to_vcf}
\alias{convert_to_vcf}
\title{convert_to_vcf Function}
\usage{
convert_to_vcf(mutations, cna, metadata)
}
\arguments{
\item{mutations}{mutations}

\item{cna}{cna}

\item{metadata}{metadata}
}
\value{
vcf object
}
\description{
convert_to_vcf Function
}
