% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fit timing of clonal peaks in cancer genome sequencing data}
\usage{
fit(
  segments,
  mutations,
  purity,
  possible_k = c("2:1", "2:2", "2:0"),
  alpha = 0.05,
  min_mutations_number = 2,
  beta_binomial = FALSE,
  beta_binomial_disp = 0.01
)
}
\arguments{
\item{segments}{A data frame containing segment information with columns \code{chr}, \code{from}, \code{to}, \code{Major}, and \code{minor}.}

\item{mutations}{A data frame containing mutation information with columns \code{chr}, \code{from}, \code{to}, \code{DP}, and \code{NV}.}

\item{purity}{A numeric value representing the tumor purity.}

\item{possible_k}{A character vector of possible karyotypes in the format "Major:minor". Default is c("2:1", "2:2", "2:0").}

\item{alpha}{A numeric value for the significance level. Default is 0.05.}

\item{min_mutations_number}{An integer specifying the minimum number of mutations required for analysis. Default is 2.}

\item{beta_binomial}{A logical value indicating whether to use the beta-binomial model. Default is FALSE.}

\item{beta_binomial_disp}{A numeric value for the beta-binomial dispersion parameter. Default is 0.01.}
}
\value{
A list containing two tibbles: \code{inference_results} and \code{summarized_results}. Returns NULL if no results are obtained.
}
\description{
This function fits the timing of clonal peaks in cancer genome sequencing data using either a beta-binomial or binomial model.
}
