% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_h.R
\name{fit_h}
\alias{fit_h}
\title{Get the fit to obtain the clocks posteriors}
\usage{
fit_h(
  x,
  max_attempts = 2,
  INIT = TRUE,
  tolerance = 1e-04,
  possible_k = c("2:1", "2:2", "2:0"),
  alpha = 0.05,
  min_mutations_number = 4,
  n_components = 0,
  initial_iter = 200,
  grad_samples = 10,
  elbo_samples = 200,
  tmp_file_path = NULL,
  cmd_version_old = FALSE
)
}
\arguments{
\item{x}{list: A CNAqc object.}

\item{max_attempts}{num: max number of repeated inference for ADVI}

\item{INIT}{logical: boolean variable to set the initialization phase to TRUE or FALSE}

\item{tolerance}{num: tolerance in the ELBO optimization procedure}

\item{possible_k}{chr: "2:1" "2:2" "2:0"}

\item{alpha}{num: (type double) confidence interval level to choose the data that fall in the expected binomial intervals}

\item{min_mutations_number}{num: (type double) minimum number of accepted mutations for a segment to be included in the inference}

\item{n_components}{number of components specified from user}

\item{initial_iter}{description}

\item{grad_samples}{description}

\item{elbo_samples}{description}

\item{tmp_file_path}{path of the directory where to save the temporary files with the info on the elbo evaluations during the VI inference}

\item{cmd_version_old}{version of cmdstanr for the draws parameter invariational method}
}
\value{
results_and_data = list(data = input_data_list, results = results, output_files_list = output_files_list)
}
\description{
Obtain the approximate posterior of the clocks for each model fit with the number of components up to k_max.
}
