% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_variational_h.R
\name{fit_variational_h}
\alias{fit_variational_h}
\title{fit_variational Function}
\usage{
fit_variational_h(
  input_data,
  max_attempts = 2,
  initialization = NULL,
  INIT = TRUE,
  initial_iter = 1000,
  grad_samples = 10,
  elbo_samples = 100,
  tolerance = 0.01
)
}
\arguments{
\item{input_data}{list: List of 7: $S: int, $N: int, $karyotype: num (0 or 1), $seg_assignment: num, $peaks:List of N of num (1:2), $NV: num, $DP: num}

\item{max_attempts}{num: max number of repeated inference for ADVI}

\item{initialization}{list: List of 4: $w: num (1:S, 1:3), $tau: num (1:K), $phi: num (1:K), $kappa: num}

\item{INIT}{logical: boolean variable to set the initialization phase to TRUE or FALSE}

\item{initial_iter}{description}

\item{grad_samples}{description}

\item{elbo_samples}{description}

\item{tolerance}{num: tolerance in the ELBO optimization procedure}
}
\value{
best_fit
}
\description{
This function performs the inference using the ADVI algorithm. Repeat inference if it fails and repeat to avoid local minima, taking the best run.
}
\examples{
fit_variational()
}
\keyword{variational}
