% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_functions.R, R/utils_h.R
\name{get_clonal_peaks}
\alias{get_clonal_peaks}
\title{get_clonal_peaks Function}
\usage{
get_clonal_peaks(k, purity)

get_clonal_peaks(k, purity)
}
\arguments{
\item{k}{karyotype}

\item{purity}{peaks}
}
\value{
get the expected VAF peaks from the karyotype and the sample purity value
}
\description{
Obtain the theoretical peaks that would be observed in a VAF spectrum.
}
\keyword{peaks}
