% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnaqc.R
\name{get_drivers}
\alias{get_drivers}
\title{Extract driver mutations from mutation data}
\usage{
get_drivers(x, chromosomes = paste0("chr", c(1:22, "X", "Y")), which = "VAF")
}
\arguments{
\item{x}{A CNAqc object containing mutation data.}

\item{chromosomes}{A character vector of chromosome names to filter for (default: 'chr1' to 'chr22', 'X', 'Y').}

\item{which}{A character string specifying whether to extract 'VAF' or 'CCF' values.}
}
\value{
A data frame of driver mutations filtered by chromosome and data type.
}
\description{
Extract driver mutations from mutation data
}
