% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_k_max_k_means.R
\name{get_k_max_k_means}
\alias{get_k_max_k_means}
\title{model_selection Function}
\usage{
get_k_max_k_means(input_data, purity, alpha = 0.05)
}
\arguments{
\item{input_data}{input data}

\item{purity}{sample purity}

\item{alpha}{description}
}
\value{
optimal_k: suggested number from c-means
}
\description{
Perform model selection among the models fit with varying number of mixture components (number of clocks).
}
\keyword{fit}
