% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_functions.R
\name{get_major}
\alias{get_major}
\title{get_major Function}
\usage{
get_major(k, get_minor = FALSE)
}
\arguments{
\item{k}{k}

\item{get_minor}{boolean value}
}
\value{
get the major or minor allele from the karyotype
}
\description{
get_major Function
}
