% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnaqc.R
\name{get_reference}
\alias{get_reference}
\title{Get reference genome coordinates}
\usage{
get_reference(ref)
}
\arguments{
\item{ref}{A character string specifying the reference genome ('hg19' or 'hg38').}
}
\value{
A data frame with chromosome coordinates for the specified reference genome.
}
\description{
Get reference genome coordinates
}
