% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_functions.R
\name{get_simulation}
\alias{get_simulation}
\title{get_simulation Function}
\usage{
get_simulation(
  taus,
  karyotypes,
  purity = 0.9,
  time_interval = 20,
  l = 1e+07,
  mu = 1e-04,
  w = 0.01,
  coverage = 100
)
}
\arguments{
\item{taus}{= a vector of doubles between 0-1}

\item{karyotypes}{= a vector of the same length of taus of strings of the type "2:1", "2:0", "2:2"}

\item{purity}{sample purity}

\item{time_interval}{time interval not in 0-1, real number}

\item{l}{length of the segment considered}

\item{mu}{mutation rate}

\item{w}{cell division rate}

\item{coverage}{average number of reads that align to a reference base}
}
\value{
all the mutations associated with the simulated CN events
}
\description{
This function allows you to obtain simulated data for copy number events.
}
