% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_functions.R
\name{get_simulation_tickTack}
\alias{get_simulation_tickTack}
\title{get_taus_karyo Function}
\usage{
get_simulation_tickTack(
  number_clocks,
  number_events,
  purity,
  coverage,
  epsilon,
  seed = 123,
  vector_karyo = c("2:0", "2:1", "2:2"),
  time_interval = 7,
  w = 0.01,
  mu = 1e-04,
  l = 2e+07
)
}
\arguments{
\item{number_clocks}{unique tau to be present in the simulation}

\item{number_events}{number of events to be simulated}

\item{purity}{sample purity}

\item{coverage}{average number of reads that align to a reference base}

\item{epsilon}{minimum distance between taus in terms of pseudo time}

\item{seed}{simulation seed}

\item{vector_karyo}{possible karyotypes to be included in the inference}

\item{time_interval}{time interval}

\item{w}{cell division rate}

\item{mu}{mutation rate}

\item{l}{length of the segments}
}
\value{
list of simulated data and x object on which to perform the inference
}
\description{
This function allows you to obtain simulated data for copy number events (tau and karyo input for get_simulation or simulate_and_fit) inputting the number of events and the proportions of taus and karyos.
}
