% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_functions.R
\name{get_taus_karyo}
\alias{get_taus_karyo}
\title{get_taus_karyo Function}
\usage{
get_taus_karyo(
  number_events,
  vector_tau,
  vector_karyo,
  weigths_tau,
  weights_karyo
)
}
\arguments{
\item{number_events}{number of events to be simulated}

\item{vector_tau}{unique tau to be present in the simulation}

\item{vector_karyo}{unique karyotype to be present in the simulation}

\item{weigths_tau}{proportion of tau to generate for each unique value}

\item{weights_karyo}{proportion of karyo to generate for each unique karyo}
}
\value{
list of simulated taus and karyo
}
\description{
This function allows you to obtain simulated data for copy number events (tau and karyo input for get_simulation or simulate_and_fit) inputting the number of events and the proportions of taus and karyos.
}
