% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnaqc.R
\name{has_driver_data}
\alias{has_driver_data}
\title{Check if mutation data has driver annotations}
\usage{
has_driver_data(x)
}
\arguments{
\item{x}{A CNAqc object containing mutation data.}
}
\value{
Logical (\code{TRUE} if driver annotation columns exist, otherwise \code{FALSE}).
}
\description{
Check if mutation data has driver annotations
}
