% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_h.R
\name{karyo_to_int}
\alias{karyo_to_int}
\title{karyo_to_int Function}
\usage{
karyo_to_int(k)
}
\arguments{
\item{k}{karyotype string}
}
\description{
This function allows you to change the string representation of the CN events into numeric values 0 - 1.
}
\examples{
karyo_to_int()
}
\keyword{karyotype}
