% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_h.R
\name{model_selection_h}
\alias{model_selection_h}
\title{model_selection Function}
\usage{
model_selection_h(results, n_components = 0)
}
\arguments{
\item{results}{list of 4: $data, $draws_and_summary, $log_lik_matrix_list and $elbo_iterations}

\item{n_components}{number of components specified from user}
}
\value{
result_model_selection: list $best_fit, $best_K, $model_selection_tibble, $entropy_list)
}
\description{
Perform model selection among the models fit with varying number of mixture components (number of clocks).
}
\keyword{fit}
