% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcawg_example}
\alias{pcawg_example}
\title{pcawg_example}
\format{
\subsection{\code{pcawg_example}}{

A list:
\describe{
\item{mutations}{A tibble with observed mutations}
\item{cna}{A tibble with observed copy number alterations}
\item{metadata}{A tibble containing sample metadata}
}
}
}
\source{
.
}
\usage{
pcawg_example
}
\description{
Pcawg example to test timing inference
}
\keyword{datasets}
