% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnaqc.R
\name{plot_cnaqc}
\alias{plot_cnaqc}
\title{Plot}
\usage{
plot_cnaqc(
  x,
  chromosomes = paste0("chr", c(1:22)),
  add_mobster = FALSE,
  max_Y_height = 6,
  cn = "absolute",
  highlight = x$most_prevalent_karyotype,
  highlight_QC = FALSE
)
}
\arguments{
\item{x}{A CNAqc object containing mutation data.}

\item{chromosomes}{A character vector of chromosome names to filter for (default: 'chr1' to 'chr22', 'X', 'Y').}

\item{add_mobster}{TRUE or FALSE}

\item{max_Y_height}{heiht plot}

\item{cn}{type of cn}

\item{highlight}{which karyotype to highlight}

\item{highlight_QC}{FALSE}
}
\value{
Plot.
}
\description{
Plot
}
