% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elbo_h.R
\name{plot_elbo_h}
\alias{plot_elbo_h}
\title{Plot elbo Function}
\usage{
plot_elbo_h(elbo_iteration)
}
\arguments{
\item{elbo_iteration}{data.frame':    #iterations until convergence of elbo obs. of  2 variables: iteration: int, elbo     : num
one of the element list elbo_iteration obtained from fit_h "results"}
}
\value{
p
}
\description{
Plot the ELBO behaviour for each fit fixing the number of cluster.
}
