% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_plot_h.R
\name{plot_model_selection_h}
\alias{plot_model_selection_h}
\title{Plot the behavior of the model selection scores vs the number of clusters}
\usage{
plot_model_selection_h(model_selection_tibble, best_K)
}
\arguments{
\item{model_selection_tibble}{a tibble with 3 scores and k_max values, one for each inference}

\item{best_K}{the integer corresponding to the best number of components}
}
\value{
model_selection_plot
}
\description{
Plot the BIC, Log Likelihood, ICL scores obtatined after the fit for each number of cluster.
}
