% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior_clocks_h.R
\name{plot_posterior_clocks_h}
\alias{plot_posterior_clocks_h}
\title{Plot posterior clocks distributions obtained from the hierarchical model fit}
\usage{
plot_posterior_clocks_h(results, K)
}
\arguments{
\item{results}{list of 4: $data, $draws_and_summary, $log_lik_matrix_list and $elbo_iterations}

\item{K}{index of inference whose results want to be plotted}
}
\value{
areas_tau
}
\description{
Plot posterior clocks distributions obtained from the hierarchical model fit
}
