% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnaqc.R
\name{plot_segments_tick_tack}
\alias{plot_segments_tick_tack}
\title{Plot segmentation results from tickTack analysis}
\usage{
plot_segments_tick_tack(x, colour_by = "clock_mean", K = 1)
}
\arguments{
\item{x}{An object containing segmentation results.}

\item{colour_by}{A character string specifying the variable to color segments by (default: 'clock_mean').}

\item{K}{Numeric, number of clusters (default: 1).}
}
\value{
A ggplot object visualizing tickTack segmentation.
}
\description{
Plot segmentation results from tickTack analysis
}
