% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnaqc.R
\name{plot_segments_tick_tack_CN}
\alias{plot_segments_tick_tack_CN}
\title{Plot copy number segments from CNAqc}
\usage{
plot_segments_tick_tack_CN(
  cnaqc_x,
  K = K,
  max_alleles = 6,
  chromosomes = paste0("chr", c(1:22))
)
}
\arguments{
\item{cnaqc_x}{A CNAqc object containing CNA data.}

\item{K}{Numeric, number of clusters.}

\item{max_alleles}{Numeric, maximum number of alleles to plot (default: 6).}

\item{chromosomes}{A character vector specifying which chromosomes to plot.}
}
\value{
A ggplot object showing CNA segments.
}
\description{
Plot copy number segments from CNAqc
}
