% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnaqc.R
\name{plot_segments_tick_tack_data}
\alias{plot_segments_tick_tack_data}
\title{Plot mutation data from tickTack analysis}
\usage{
plot_segments_tick_tack_data(x, colour_by = "clock_mean", K = K)
}
\arguments{
\item{x}{An object containing mutation data.}

\item{colour_by}{A character string specifying the variable to color points by (default: 'clock_mean').}

\item{K}{Numeric, number of clusters.}
}
\value{
A ggplot object showing mutation data.
}
\description{
Plot mutation data from tickTack analysis
}
