% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timing_h.R
\name{plot_timing_h}
\alias{plot_timing_h}
\title{plot_timing_h Function}
\usage{
plot_timing_h(
  results,
  K,
  colour_by = "karyotype",
  split_contiguous_segments = TRUE
)
}
\arguments{
\item{results}{list(data = input_data_list, results = results, output_files_list = output_files_list)}

\item{K}{mun: number of clocks}

\item{colour_by}{chr: default =  "karyotype"}

\item{split_contiguous_segments}{option to plot segments' setalarion lines}
}
\value{
p : plot of inference results with credibility intervals in the chromosome absolute positions
}
\description{
This function obtains the list of input data to be used in the stan model.
}
\keyword{plot}
