% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cnaqc.R
\name{plot_vaf}
\alias{plot_vaf}
\title{Plot variant allele frequency (VAF) distributions}
\usage{
plot_vaf(x, K)
}
\arguments{
\item{x}{An object containing mutation and segmentation data.}

\item{K}{Numeric, number of clusters.}
}
\value{
A ggplot object visualizing VAF distributions.
}
\description{
Plot variant allele frequency (VAF) distributions
}
