% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_h.R
\name{predict_best_number_of_clusters}
\alias{predict_best_number_of_clusters}
\title{Model Saelection.}
\usage{
predict_best_number_of_clusters(fit, purity, coverage)
}
\arguments{
\item{fit}{results from tickTack_h inference.}

\item{purity}{sample purity}

\item{coverage}{sample coverage}
}
\value{
A tibble with the data.
}
\description{
Perform model selection data preparation.
}
