% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input_data_h.R
\name{prepare_input_data}
\alias{prepare_input_data}
\title{prepare_input_data Function}
\usage{
prepare_input_data(
  mutations,
  segments,
  purity,
  possible_k = c("2:1", "2:2", "2:0"),
  alpha = 0.05,
  min_mutations_number = 2
)
}
\arguments{
\item{mutations}{list: output from CNAqc Mutations(x), where x is a CNAqc object}

\item{purity}{num: (type = double) sample purity between 0 and 1}

\item{possible_k}{chr: "2:1" "2:2" "2:0"}

\item{alpha}{num: (type double) confidence interval level to choose the data that fall in the expected binomial intervals}

\item{min_mutations_number}{num: (type double) minimum number of accepted mutations for a segment to be included in the inference}

\item{cna}{list: output from CNAqc CNA(x), where x is a CNAqc object}
}
\value{
accepted_data: list $input_data:list: List of 7: $S: int, $N: int, $karyotype: num (0 or 1), $seg_assignment: num, $peaks:List of N of num \link{1:2}, $NV: num, $DP: num
$accepted_cna: tibble \link{S × 5} (S3: tbl_df/tbl/data.frame): $segment_original_indx: int, $segment_name: chr, $segment_id: num, $karyotype: chr, $chr: chr
}
\description{
This function obtains the list of input data to be used in the stan model.
}
\examples{
prepare_input_data()
}
\keyword{input}
