% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_h.R
\name{process_file_prediction}
\alias{process_file_prediction}
\title{Model Selection function.}
\usage{
process_file_prediction(fit)
}
\arguments{
\item{fit}{results from tickTack_h inference.}
}
\value{
A tibble with the data.
}
\description{
Perform model selection data preparation.
}
