% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_functions.R
\name{simulate_mutations}
\alias{simulate_mutations}
\title{simulate_mutations Function}
\usage{
simulate_mutations(
  karyotype,
  time_interval,
  tau,
  l,
  mu,
  w,
  segment_id = "segment_id"
)
}
\arguments{
\item{karyotype}{karyotype}

\item{time_interval}{time}

\item{tau}{tau generated}

\item{l}{length of the segment considered}

\item{mu}{mutation rate}

\item{w}{cell division rate}

\item{segment_id}{segment id}
}
\value{
df with all the mutations associated with a simulated CN event
}
\description{
This function allows you to obtain sample mutations to use for CN Timing inference.
}
