% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_h.R
\name{get_clonal_peaks}
\alias{get_clonal_peaks}
\title{get_clonal_peaks Function}
\usage{
get_clonal_peaks(k, purity)
}
\arguments{
\item{k}{karyotype}

\item{purity}{peaks}
}
\description{
Obtain the theoretical peaks that would be observed in a VAF spectrum.
}
\examples{
get_clonal_peaks()
}
\keyword{peaks}
