% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection_h.R
\name{model_selection_h}
\alias{model_selection_h}
\title{model_selection Function}
\usage{
model_selection_h(
  data,
  draws_and_summary,
  log_lik_matrix_list,
  elbo_iterations
)
}
\arguments{
\item{data}{list of k_max of lists: $input_data:list: List of 7: $S: int, $N: int, $karyotype: num (0 or 1), $seg_assignment: num, $peaks:List of N of num \link{1:2}, $NV: num, $DP: num}

\item{results}{list of lenght k_max of results form the variational method with the summary statistics of the draws from the approximate posterior}
}
\value{
result_model_selection: list $best_fit, $best_K, $model_selection_tibble, $entropy_list)
}
\description{
Perform model selection among the models fit with varying number of mixture components (number of clocks).
}
\keyword{fit}
