% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_inference_h.R
\name{plot_inference_h}
\alias{plot_inference_h}
\title{plot_inference_h Function}
\usage{
plot_inference_h(
  results,
  K,
  colour_by = "karyotype",
  split_contiguous_segments = TRUE
)
}
\arguments{
\item{results}{list(data = input_data_list, results = results, output_files_list = output_files_list)}

\item{K}{mun: number of clocks}

\item{colour_by}{chr: default =  "karyotype"}

\item{x_segments}{tibble((S3: tbl_df/tbl/data.frame) chr, from, to, Major, minor, total_cn data$cna)}

\item{input_data}{List of 7: $S: int, $N: int, $karyotype: num (0 or 1), $seg_assignment: num, $peaks:List of N of num (1:2), $NV: num, $DP: num}
}
\value{
p : plot of inference results with credibility intervals in the chromosome absolute positions
}
\description{
This function obtains the list of input data to be used in the stan model.
}
\keyword{plot}
