% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tickTack-package.R
\docType{package}
\name{tickTack-package}
\alias{tickTack}
\alias{tickTack-package}
\title{tickTack: Timing Inference for Copy number alterations: Tracking Alterations in Cancer Karyotypes}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
